massage Tabs instead
Because the ->{data} of a LifecycleObj pulls from the global lifecycle cache,
altering it alters it for the process lifetime.  Under a standalone server, if
I load a ticket with a user with this preference set to "Respond", it _also_
affects all other users who don't have the preference set. chmrr++